Reorganize Flutter app into feature folders
