+++ crm_retail_app/lib/features/dashboard/metric_detail_screen.dart
