+++ crm_retail_app/lib/features/dashboard/store_detail_screen.dart
