+++ crm_retail_app/lib/features/dashboard/tabs/inventory_tab.dart
